Fix RealCanvas coordinate mapping
Until now, given real upper bounds (bx, by), with underlying
canvas size (sx, sy), coordinates bx and by would map directly
to sx and sy, thus out of the underlying discrete canvas.

This created some problems, for example a shape sized 1.0 on
one of the dimensions would not fit a RealCanvas of size 1.0 on
the same dimension.

With this change, coordinates equal to the upper bounds will
map to the last pixel of the canvas, thus bx -> (sx - 1) and
by -> (sy - 1).